chore: refactor and deduplicate types tests
